requirements/dev.txt
# Everything the developer needs in addition to the production requirements

-r prod.txt

# Testing
pytest

python-dotenv
